refactor: AnalyticsPanel 資料來源改走 Session 與 DataFacade
- 移除原本直接操作 game_log 的邏輯
- 改用 Session 取得資料並統一資料流
- 介面成功顯示資料分析結果
- 部份資料正確顯示
- 還有一些錯誤要處理
