+++ ui/main_window.py
@@ -3,7 +3,7 @@ from PySide6.QtWidgets import (
     QTabWidget
 )
 from ui.pages.betting_page import BettingPage
-#from ui.pages.analytics_page import AnalyticsPage
+from ui.pages.analytics_page import AnalyticsPage
 #from ui.pages.training_page import TrainingPage
 
 
@@ -14,7 +14,7 @@ class MainWindow(QMainWindow):
 
         self.tabs = QTabWidget()
         self.tabs.addTab(BettingPage(), "🎮 下注頁")
-        #self.tabs.addTab(AnalyticsPage(), "📈 資料分析")
+        self.tabs.addTab(AnalyticsPage(), "📈 資料分析")
         #self.tabs.addTab(TrainingPage(), "🧠 模型訓練")
         self.setCentralWidget(self.tabs)
